* revert a merge error

* try to fix the sound problem until 3 hours playing

* Update Makefile

* Update build.yml

* try to fix duplicate guid problem

by using this fix https://github.com/ioquake/ioq3/pull/485/commits/bd48cdbd5eae0924e9d71b5736134f9581182953 provided by rdnt

* im dumb

* come back to old sdl for windows since it will be disabled by default

* Update build.yml

* 32bit build for windows are built with mingw64 so the release name must be mingw64-x86

* Update quake3.png

* Update build.yml

* fix duplicate guid problem

* MSVC2005: avoid compiler syntax error, backported from jbruns/Quake3e

* Update build.yml

* Update build.yml

* Update Makefile

* Update README.md

* Update q_shared.h

* Update qcommon.h

* update logo

* Update make-macosx-app.sh

---------

+++ code/client/snd_mix.c
@@ -344,7 +344,7 @@ static void S_TransferPaintBuffer( int endtime, byte *buffer )
 		p = (int *) paintbuffer;
 		count = (endtime - s_paintedtime) * dma.channels;
 		out_mask = dma.samples - 1; 
-		out_idx = ( s_paintedtime * dma.channels ) & out_mask;
+		out_idx = ( (unsigned int)s_paintedtime * dma.channels ) & out_mask;
 		step = 3 - dma.channels;
 
 		if ( dma.samplebits == 32 && dma.isfloat )
